refactoring(racingGame): game 하위 클래스들을 domain, view 패키지로 분리
