+++ src/main/java/game/domain/MovePolicy.java
@@ -1,4 +1,4 @@
-package game;
+package game.domain;
 
 public interface MovePolicy {
     public boolean isMovable();
