+++ src/test/java/game/domain/RandomMovePolicyTest.java
@@ -1,5 +1,7 @@
-package game;
+package game.domain;
 
+import game.domain.MovePolicy;
+import game.domain.RandomMovePolicy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -9,9 +11,7 @@ import java.lang.reflect.Field;
 import java.util.Random;
 import java.util.stream.Stream;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.assertj.core.api.Assertions.*;
-import static org.mockito.Mockito.*;
 
 class RandomMovePolicyTest {
